Implement Lazy loading tree + adding message to delete button

+++ web/src/main/java/com/engagepoint/labs/web/controllers/ActionBean.java
@@ -6,10 +6,12 @@ import com.engagepoint.labs.core.models.FSObject;
 import com.engagepoint.labs.core.service.CMISService;
 import com.engagepoint.labs.core.service.CMISServiceImpl;
 import org.primefaces.context.RequestContext;
+import org.primefaces.model.TreeNode;
 
 import javax.faces.application.FacesMessage;
 import javax.faces.bean.ApplicationScoped;
 import javax.faces.bean.ManagedBean;
+import javax.faces.bean.ManagedProperty;
 import javax.faces.component.UIComponent;
 import javax.faces.context.FacesContext;
 import java.io.Serializable;
@@ -22,8 +24,8 @@ import java.util.logging.Logger;
 
 @ManagedBean(name = "action")
 @ApplicationScoped
-public class ActionBean implements Serializable {
 
+public class ActionBean implements Serializable {
     private String type;
     private String reqEx;
     private String name;
@@ -31,13 +33,17 @@ public class ActionBean implements Serializable {
     private boolean deleteAllTree = false;
     private UIComponent renamecomponent;
     private UIComponent createcomponent;
+    private UIComponent deletecomponent;
     private static Logger logger;
     private final CMISService CMISservice;
 
+    @ManagedProperty(value = "#{treeBean}")
+    private TreeBean treeBean;
+
     /**
      * Handling exception and create a message to show user om dialog page  and log the exception
-     * method fail validation and skip all the subsequent phases and go to render response
-     * phase to avoid closing dialog with the client
+     * method fail validation and skip all the subsequent phases and go to render response phase
+     * to avoid closing dialog with the client
      *
      * @param ex        Exception that is thown from service layer
      * @param component Component to which error is binding
@@ -68,11 +74,13 @@ public class ActionBean implements Serializable {
     public void createFolder(FSFolder parent) {
         try {
             CMISservice.createFolder(parent, name);
+            treeBean.updateTree(treeBean.getSelectedNode());
         } catch (Exception ex) {
             catchException(ex, createcomponent);
         }
         this.name = "";
         this.type = "";
+
     }
 
     /**
@@ -85,13 +93,17 @@ public class ActionBean implements Serializable {
         try {
             if (fsObject instanceof FSFolder) {
                 CMISservice.renameFolder((FSFolder) fsObject, newName);
+
             } else if (fsObject instanceof FSFile) {
                 CMISservice.renameFile((FSFile) fsObject, newName);
             }
         } catch (Exception ex) {
             catchException(ex,renamecomponent);
         }
+        treeBean.updateTree(treeBean.getSelectedNode().getParent());
+        treeBean.getSelectedNode().setExpanded(true);
         this.newName = "";
+        
     }
 
     /**
@@ -105,8 +117,13 @@ public class ActionBean implements Serializable {
             deleteAllTree = false;
             CMISservice.deleteAllTree(folder);
         } else {
-            CMISservice.deleteFolder(folder);
+            if (!CMISservice.hasChildren(folder)) CMISservice.deleteFolder(folder);
+            else {
+                Exception ex = new Exception("Folder has subfolders! To remove folder select checkbox ") ;
+                catchException(ex, deletecomponent);
+            }
         }
+        treeBean.updateTree(treeBean.getSelectedNode().getParent());
     }
 
     public String getName() {
@@ -163,5 +180,20 @@ public class ActionBean implements Serializable {
 
     public void setReqEx(String reqEx) {
         this.reqEx = reqEx;
+    }
+    public TreeBean getTreeBean() {
+        return treeBean;
+    }
+
+    public void setTreeBean(TreeBean treeBean) {
+        this.treeBean = treeBean;
+    }
+
+    public UIComponent getDeletecomponent() {
+        return deletecomponent;
+    }
+
+    public void setDeletecomponent(UIComponent deletecomponent) {
+        this.deletecomponent = deletecomponent;
     }
 }
